Move UI init code out of modules.
