+++ package.json
@@ -2,7 +2,7 @@
   "name": "blocklyprop-solo",
   "version": "1.5.6",
   "description": "A simplified implementation of Google's Blockly project configured to support Parallax robots and sensors.",
-  "main": "editor.js",
+  "main": "load_images.js",
   "scripts": {
     "compile:sass": "sass sass/main.scss dist/main.css -w",
     "dev": "webpack --config webpack/dev.config.js --node-env development --watch",
