+++ sass/main.scss
@@ -347,6 +347,10 @@ ul > hr {
   box-shadow: inset 0 1px 1px rgba(0,0,0,0.0);
 }
 
+#right-menus {
+  float:right;
+  padding-right: 15px;
+}
 
 #comPort {
   display: inline-block;
