+++ src/templates/index.html
@@ -67,8 +67,9 @@
     <div id="editor">
         <!-- Branding logo -->
         <div id="branding">
-            <span id="nav-logo" class="url-prefix"></span>
-<!--            <a id="nav-logo" href="/" class="url-prefix"></a>-->
+            <span id="nav-logo" class="url-prefix">
+                BlocklyProp<br><strong>Solo</strong>
+            </span>
         </div>
 
         <div id="noScript">
@@ -77,7 +78,6 @@
             </noscript>
         </div>
 
-
         <div id="editor-status">
             <!-- Above toolbar -->
             <div style="overflow:auto;">
@@ -85,13 +85,16 @@
                 <!-- BlocklyProp Client or Loader UI -->
                 <div id="client-status-alerts" style="float: left;">
                     <span style="padding-left: 10px; line-height: 30px;">
-                        <span id="client-unavailable" class="bp-client-danger">
-                            <a class="client-unavailable-link" data-toggle="modal" data-target="#client-download-modal"
-                                href="#">
+                        <span id="client-loading" class="bp-client-available">Loading...</span>
+
+                        <span id="client-unavailable" class="bp-client-danger hidden">
+                            <a class="client-unavailable-link" data-toggle="modal"
+                               data-target="#client-download-modal" href="#">
                                 <span class="bpIcon" data-icon="dangerTriangle">!!</span>
                                 <span class="keyed-lang-string" data-key="editor_client_not-available"></span>
                             </a>
                         </span>
+
                         <span id="client-available" class="bp-client-available keyed-lang-string hidden"
                             data-key="editor_client_available">
                         </span>
@@ -150,41 +153,39 @@
                 </div>
 
                 <!-- Right Toolbar buttons and menu -->
-                <div id="right-menus" style="float:right; padding-right: 15px;">
+                <div id="right-menus">
                     <!-- View C source code -->
-                    <a class="btn-view-blocks" id="btn-view-propc" style="display: none;">
-                            <span class="bpIcon" data-icon="eyeWhite">&#x1F441;</span>
-                            &nbsp;
-                            <span class="keyed-lang-string" data-key="menu_code"></span>
-                        </a>
+                    <a class="btn-view-blocks" id="btn-view-propc" style="display: inline;">
+                        <span class="bpIcon" data-icon="eyeWhite">&#x1F441;&nbsp;Code</span>
+                        <span>Code</span>
+                    </a>
 
                     <!-- View project blocks layout -->
                     <a class="btn-view-blocks" id="btn-view-blocks" style="display: none;">
-                            <span class="bpIcon" data-icon="eyeWhite">&#x1F441;</span>
-                            &nbsp;
-                            <span class="keyed-lang-string" data-key="menu_blocks"></span>
+                        <span class="bpIcon" data-icon="eyeWhite">&#x1F441;&nbsp;</span>
+                        <span>Blocks</span>
                         </a>
 
                     <!-- View project XML code -->
                     <a class="btn-view-blocks" id="btn-view-xml" style="display: none;">
-                            <span class="bpIcon" data-icon="eyeWhite">&#x1F441;</span>
-                            &nbsp;&nbsp;
-                            <span class="keyed-lang-string" data-key="editor_view_xml"></span>
-                        </a>
-
-                    <!-- New Project button -->
-                    <a id="new-project-button" class="demo-function btn-view-blocks" style="display: inline-block;" data-displayas="inline-block">
-                        <span class="keyed-lang-string" data-key="editor_new_button"></span>
+                        <span class="bpIcon" data-icon="eyeWhite">&#x1F441;&nbsp;</span>
+                        <span>XML</span>
                     </a>
 
+                    <!-- New Project button -->
+                    <a id="new-project-button" class="demo-function btn-view-blocks"
+                       style="display: inline-block;" data-displayas="inline-block">New</a>
+
                     <!-- Open Project button -->
-                    <a id="open-project-button" class="demo-function btn-view-blocks" style="display: inline-block;" data-displayas="inline-block">
-                        <span class="keyed-lang-string" data-key="editor_open_button"></span>
+                    <a id="open-project-button" class="demo-function btn-view-blocks"
+                       style="display: inline-block;" data-displayas="inline-block">
+                        <span>Open</span>
                     </a>
 
                     <!-- Save project button -->
-                    <a class="demo-function btn-view-blocks" id="save-project" style="display: inline-block;" data-displayas="inline-block">
-                        <span class="keyed-lang-string" data-key="editor_save"></span>
+                    <a id="save-project" class="demo-function btn-view-blocks"
+                       style="display: inline-block;" data-displayas="inline-block">
+                        <span>Save</span>
                     </a>
 
                     <!-- Drop-down hamburger menu -->
@@ -929,6 +930,9 @@
             </div>
         </div>
     </div>
-
+    <script defer>
+      console.log('Loading images...');
+    </script>
+    <script src="load_images.js"></script>
 </body>
 </html>
