+++ webpack/dev.config.js
@@ -117,6 +117,7 @@ module.exports = (opts) => {
           {from: path.resolve(__dirname, blocklyMedia), to: path.resolve(__dirname, `${targetPath}/media`)},
           {from: './src/images', to: path.resolve(__dirname, `${targetPath}/images`)},
           {from: './sass/main.css', to: path.resolve(__dirname, targetPath)},
+          {from: './src/load_images.js', to: path.resolve(__dirname, targetPath)},
         ]
       })
     ],
